@@ -1,6 +1,5 @@
-const F = require('lodash/fp');
-const isEmpty = require('lodash/isEmpty');
-const Yup = require('./yup');
+import F from 'lodash/fp';
+import Yup from './yup';
 
 function and (...args) {
     for (let i = 0; i < args.length; i++) {
@@ -121,9 +120,9 @@ function calcValidConstraints ({ deviceSize, constraints, skipValidation }) {
             ? when(deviceSize, F.pick(['width', 'height'], constraint))
             : when;
         let doesPassConstraint;
-        if (!isEmpty(constraint.whenAny)) {
+        if (!F.isEmpty(constraint.whenAny)) {
             doesPassConstraint = or(...constraint.whenAny.map(mapWhen));
-        } else if (!isEmpty(constraint.whenAll)) {
+        } else if (!F.isEmpty(constraint.whenAll)) {
             doesPassConstraint = and(...constraint.whenAll.map(mapWhen));
         } else {
             throw new Error(`No 'whenAny' or 'whenAll' param found!`);
