@@ -1,4 +1,4 @@
-const Yup = require('yup');
+import Yup from 'yup';
 
 class FunctionSchema extends Yup.mixed {
     _typeCheck(v) {
@@ -47,7 +47,7 @@ const createSchema = (type) => (...args) => {
     return this;
 };
 
-module.exports = Object.assign({}, Yup, {
+export default Object.assign({}, Yup, {
     func: createSchema(FunctionSchema),
     object: createSchema(ObjectSchema),
     alternatives: createSchema(AlternativesSchema),
